refactor: split serenity's command
